--- client/src/components/Toolbox/styles/Toolbox.scss
@@ -1,137 +1,0 @@
-.toolbox-container {
-  background: url("../../../assets/images/background.png");
-  background-position: center;
-  height: 100vh;
-  padding: 124px 213px;
-  width: 100%;
-  background-size: cover;
-  background-repeat: no-repeat;
-
-
-  .categories {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    gap: 16px;
-    .category {
-      text-align: center;
-      width: 100%;
-      color: #854d15;
-      border: 1px solid #d58f49;
-      text-align: center;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      column-gap: 20px;
-      font-size: 24px;
-      height: 78px;
-      font-weight: normal;
-      background-color: #dac29c;
-      border : 2px solid #dac29c;
-      box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.5);
-
-      cursor: pointer;
-
-      &:hover {
-        background-color: #e0e0e0; 
-      }
-
-      &.active {
-        background-color: rgba(218, 194, 156, 0.514);
-        font-weight: normal;
-      }
-    }
-  }
-  .board {
-    margin-top: 30px;
-    background-color: #DAC29C;
-    padding: 20px 60px;
-
-    .header{
-      display:  flex;
-      align-items: center;
-      justify-content: space-between;
-      padding-bottom: 50px;
-      border-bottom: 1px solid #854D15;
-      p{
-        font-size: 24px;
-        color: #854D15;
-      }
-    }
-
-    .options {
-      display: grid;
-      padding: 12px 0;
-      grid-template-columns: repeat(3, 1fr);
-      gap: 78px;
-    }
-  }
-
-
-  .active-category {
-    display: flex;
-    align-items: center;
-    padding: 10px 12px;
-    background-color: #230056;
-    color: #f5f5f5;
-    border-radius: 0 0 12px 12px;
-    gap: 4px;
-    h3 {
-      font-weight: 400;
-      font-size: 16px;
-    }
-    h5 {
-      font-weight: bold;
-      font-size: 16px;
-      padding: 0 4px;
-      border-left: rgba(45, 204, 253, 0.48) 2px solid;
-    }
-  }
-
-  .bottom-buttons {
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 21px 0px;
-    .new-game{
-      width: 293px;
-      height: 88px;
-      background: #854D15;
-      font-size: 24px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 20px;
-      color: #DAC29C;
-      border: 1px solid #D58F49;
-    }
-
-    .toolbox-settings{
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: 10px;
-     button{
-      width: 146px;
-      
-      height: 54px;
-      color: #854D15;
-      background: #DAC29C;
-      font-size: 16px;
-      border: 1px solid #854D15;
-     }
-    }
-  }
-
-  .dice-options {
-    display: grid;
-    padding: 12px 0;
-    
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 20px;
-    img {
-      height: 70px;
-      width: 57px;
-    }
-  }
-}
