+++ client/src/styles/toolbox/Toolbox.scss
@@ -1,0 +1,260 @@
+.toolbox-container {
+  background: url("../../assets/images/background.png");
+  background-position: center;
+  height: 100vh;
+  padding: 124px 213px;
+  width: 100%;
+  background-size: cover;
+  background-repeat: no-repeat;
+
+  .categories {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 16px;
+
+    .category {
+      text-align: center;
+      width: 100%;
+      color: #854d15;
+      border: 1px solid #d58f49;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      column-gap: 20px;
+      font-size: 24px;
+      height: 78px;
+      font-weight: normal;
+      background-color: #dac29c;
+      border: 1px solid #854d15;
+      box-shadow: 0px 4px 2px 0px rgba(0, 0, 0, 0.25);
+      cursor: pointer;
+
+      &:hover {
+        background-color: #e0e0e0;
+      }
+
+      &.active {
+        background-color: rgba(218, 194, 156, 0.514);
+        font-weight: normal;
+      }
+    }
+  }
+
+  .board {
+    margin-top: 30px;
+    background-color: #dac29c;
+    padding: 20px 60px;
+
+    .header {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      padding-bottom: 50px;
+      border-bottom: 1px solid #854d15;
+
+      p {
+        font-size: 24px;
+        color: #854d15;
+      }
+    }
+
+    .options {
+      display: grid;
+      padding: 12px 0;
+      grid-template-columns: repeat(3, 1fr);
+      gap: 78px;
+    }
+  }
+
+  .active-category {
+    display: flex;
+    align-items: center;
+    padding: 10px 12px;
+    background-color: #230056;
+    color: #f5f5f5;
+    border-radius: 0 0 12px 12px;
+    gap: 4px;
+
+    h3 {
+      font-weight: 400;
+      font-size: 16px;
+    }
+
+    h5 {
+      font-weight: bold;
+      font-size: 16px;
+      padding: 0 4px;
+      border-left: rgba(45, 204, 253, 0.48) 2px solid;
+    }
+  }
+
+  .bottom-buttons {
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 21px 0;
+
+    .new-game {
+      width: 293px;
+      height: 88px;
+      background: #854d15;
+      font-size: 24px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      gap: 20px;
+      color: #dac29c;
+      border: 1px solid #d58f49;
+    }
+
+    .toolbox-settings {
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+      gap: 10px;
+
+      button {
+        width: 146px;
+        height: 54px;
+        color: #854d15;
+        background: #dac29c;
+        font-size: 16px;
+        border: 1px solid #854d15;
+      }
+    }
+  }
+
+  .dice-options {
+    display: grid;
+    padding: 12px 0;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+
+    img {
+      height: 70px;
+      width: 57px;
+    }
+  }
+
+  /* Responsive styles */
+  @media (max-width: 1200px) {
+    padding: 100px 150px;
+
+    .categories {
+      grid-template-columns: repeat(3, 1fr);
+    }
+
+    .board {
+      padding: 20px 40px;
+
+      .options {
+        gap: 50px;
+      }
+    }
+
+    .bottom-buttons {
+      .new-game {
+        width: 220px;
+        height: 70px;
+        font-size: 20px;
+      }
+
+      .toolbox-settings button {
+        width: 120px;
+        height: 50px;
+        font-size: 14px;
+      }
+    }
+  }
+
+  @media (max-width: 992px) {
+    padding: 80px 100px;
+
+    .categories {
+      grid-template-columns: repeat(2, 1fr);
+    }
+
+    .board {
+      padding: 20px 30px;
+
+      .options {
+        grid-template-columns: repeat(2, 1fr);
+        gap: 40px;
+      }
+    }
+
+    .bottom-buttons {
+      flex-direction: column;
+      align-items: stretch;
+
+      .new-game {
+        width: 100%;
+        margin-bottom: 10px;
+      }
+
+      .toolbox-settings button {
+        width: 100%;
+      }
+    }
+  }
+
+  @media (max-width: 768px) {
+    padding: 60px 50px;
+
+    .categories {
+      grid-template-columns: 1fr;
+    }
+
+    .board {
+      padding: 20px;
+
+      .options {
+        grid-template-columns: 1fr;
+      }
+    }
+
+    .bottom-buttons {
+      .new-game,
+      .toolbox-settings button {
+        width: 100%;
+        font-size: 16px;
+      }
+    }
+  }
+
+  @media (max-width: 576px) {
+    padding: 40px 20px;
+
+    .categories .category {
+      font-size: 18px;
+      height: 60px;
+    }
+
+    .board {
+      padding: 15px;
+
+      .header p {
+        font-size: 18px;
+      }
+
+      .options {
+        gap: 20px;
+      }
+    }
+
+    .bottom-buttons {
+      padding: 15px 0;
+
+      .new-game {
+        height: 60px;
+        font-size: 18px;
+      }
+
+      .toolbox-settings button {
+        height: 40px;
+        font-size: 14px;
+      }
+    }
+  }
+}
